discard: Move selected change discarding
The discard command handles selected hunk orchestration and structural
selected-change restoration in the same module.

The project lacks a command-side support boundary for renames, deletions,
and gitlinks in discard paths. That makes the command entry point own index
updates, worktree removal, and batch submodule pointer removal.

This commit continues separating structural selected-change handling by
moving discard restoration helpers into
commands.selection.selected_change_discarding and updating discard.py to
delegate to them.

The next commit will add architecture coverage for the discard boundary.
